Add: Clojure CLI tools alias to run project from Spacemacs / Emacs
The project fails to run when using `cider-jack-in-cljs` as it cannot find the
figwheel-main namespace.  This is because CIDER is not being called with the
-A:fig alias, which has a configuration to include figwheel-main as a
dependency.

Calling `SPC u C-c M-J` will display an editable prompt for Cider jack-in, to
which you can add the -A:fig option just after the /usr/local/bin/clojure
executable name.

Rather than edit the cider jack-in command options each time, a local
configuration has been created to set the cider-clojure-clj-global-options to
use the fig alias

((clojure-mode . ((cider-clojure-cli-global-options . "-A:fig"))))

The project buffer(s) need to be closed and re-opened before the configuration
in .dir-locals.el is read into Emacs.  Once the buffer is loaded again, running
`cider-jack-in-cljs` works perfectly.